Clip logprobs to -100 to avoid JSON encoder errors
Models can return a -inf logprob for tokens that they are certain are not correct, e.g. if `top_logprobs` is larger than the possible tokens allowed in a position in a structured output.

The JSON spec doesn't allow inf or NaN values for number fields, so to avoid an encoder error, this change clips logprobs to a minimum of -100. This is the same behaviour as OpenAI.